# swagger.yaml
openapi: 3.0.0
info:
  title: Example API
  description: A simple API to demonstrate Swagger UI with CRUD operations
  version: 1.0.0
servers:
  - url: http://localhost:3000
paths:
  /:
    get:
      summary: "Hello World"
      description: "Returns a simple hello world message"
      responses:
        '200':
          description: "A hello world response"
          content:
            text/plain:
              schema:
                type: string
                example: "Hello world!"
  /items:
    get:
      summary: "Get all items"
      description: "Retrieve a list of all items"
      responses:
        '200':
          description: "A list of items"
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                    name:
                      type: string
                    description:
                      type: string
    post:
      summary: "Create a new item"
      description: "Add a new item to the database"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
      responses:
        '201':
          description: "Item created successfully"
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  name:
                    type: string
                  description:
                    type: string
  /items/{id}:
    get:
      summary: "Get an item by ID"
      description: "Retrieve the details of a specific item by its ID"
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: "Item details"
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  name:
                    type: string
                  description:
                    type: string
        '404':
          description: "Item not found"
    put:
      summary: "Update an item"
      description: "Update the name and description of an item"
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
      responses:
        '200':
          description: "Item updated successfully"
        '404':
          description: "Item not found"
    delete:
      summary: "Delete an item"
      description: "Remove an item from the database"
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: "Item deleted successfully"
        '404':
          description: "Item not found"